DMARacer changes
Patches by:
* Raphael Isemann <teemperor@gmail.com>
* Victor Duta <duta.victor.marin@gmail.com>
* Jakob Koschel <jakobkoschel@aim.com>

llvm: Converts asm macros into equivalent C code

+++ arch/x86/kernel/dumpstack_64.c
@@ -72,7 +72,7 @@ struct estack_pages {
 	[PFN_DOWN(CEA_ESTACK_OFFS(st)) ...				\
 	 PFN_DOWN(CEA_ESTACK_OFFS(st) + CEA_ESTACK_SIZE(st) - 1)] = {	\
 		.offs	= CEA_ESTACK_OFFS(st),				\
-		.size	= CEA_ESTACK_SIZE(st),				\
+		.size	= (u16)CEA_ESTACK_SIZE(st),				\
 		.type	= STACK_TYPE_EXCEPTION + ESTACK_ ##st, }
 
 /*
